backend/listener/requirements.txt
python-socketio 
requests